Move `install` helper functions to own module
